Tar i bruk Nullable-pattern for Opphold
Vi ser at Opphold kan være null i produksjon, og håndterer dette ved å implementere Nullable-patternet.
Da får vi enkelt målinger på disse tilfellene også, uten å endre mye av koden rundt.
