+++ src/main/java/no/nav/fo/veilarbregistrering/bruker/Person.java
@@ -1,9 +1,5 @@
 package no.nav.fo.veilarbregistrering.bruker;
 
-import no.nav.fo.veilarbregistrering.metrics.Metric;
-
-import java.time.LocalDate;
-
 public class Person {
 
     private final Opphold opphold;
@@ -19,7 +15,7 @@ public class Person {
     }
 
     public Opphold getOpphold() {
-        return opphold;
+        return opphold != null ? opphold : Opphold.nullable();
     }
 
     public Statsborgerskap getStatsborgerskap() {
@@ -34,103 +30,4 @@ public class Person {
                 '}';
     }
 
-    public static class Opphold implements Metric {
-
-        private final Oppholdstype type;
-        private final Periode periode;
-
-        public static Opphold of(Oppholdstype type, Periode periode) {
-            return new Opphold(type, periode);
-        }
-
-        private Opphold(Oppholdstype type, Periode periode) {
-            this.type = type;
-            this.periode = periode;
-        }
-
-        public Oppholdstype getType() {
-            return type;
-        }
-
-        @Override
-        public String fieldName() {
-            return "oppholdstype";
-        }
-
-        @Override
-        public Object value() {
-            return type.toString();
-        }
-
-        @Override
-        public String toString() {
-            return "Opphold{" +
-                    "type=" + type +
-                    ", periode=" + periode +
-                    '}';
-        }
-    }
-
-    public enum Oppholdstype {
-        MIDLERTIDIG,
-        PERMANENT,
-        OPPLYSNING_MANGLER
-    }
-
-    public static class Statsborgerskap implements Metric {
-
-        private final String statsborgerskap;
-        private final Periode periode;
-
-        public static Statsborgerskap of(String statsborgerskap, Periode periode) {
-            return new Statsborgerskap(statsborgerskap, periode);
-        }
-        private Statsborgerskap(String statsborgerskap, Periode periode) {
-            this.statsborgerskap = statsborgerskap;
-            this.periode = periode;
-        }
-
-        public String getStatsborgerskap() {
-            return statsborgerskap;
-        }
-
-        @Override
-        public String toString() {
-            return "Statsborgerskap{" +
-                    "statsborgerskap='" + statsborgerskap + '\'' +
-                    ", periode=" + periode +
-                    '}';
-        }
-
-        @Override
-        public String fieldName() {
-            return "statsborgerskap";
-        }
-
-        @Override
-        public Object value() {
-            return statsborgerskap;
-        }
-    }
-    public static class Periode {
-
-        private final LocalDate fra;
-        private final LocalDate til;
-
-        public static Periode of(LocalDate fra, LocalDate til) {
-            return new Periode(fra, til);
-        }
-        private Periode(LocalDate fra, LocalDate til) {
-            this.fra = fra;
-            this.til = til;
-        }
-
-        @Override
-        public String toString() {
-            return "Periode{" +
-                    "fra=" + fra +
-                    ", til=" + til +
-                    '}';
-        }
-    }
 }
